size_t was not defined here

+++ src/core/platform.h
@@ -99,7 +99,7 @@ typedef SSIZE_T ssize_t;
 
 /* _countof helper */
 #if !defined (_countof)
-template <typename Type, size_t Size>
+template <typename Type, unsigned Size>
 char(&__countof_helper(Type(&_Array)[Size]))[Size];
 #define _countof(_Array) (sizeof(__countof_helper(_Array)) + 0)
 #endif  /* !defined (_countof) */
